Create README.md
Mage
   	Game Design Document

               	 
Revision: 0.0.1 
 
Robson Allein F. Miguelles

 
Project Description
 
This game design document describes the details for a multi-platform touch based 2D adventure game with original story and characters. The game plays like other match-3 games but introduces some innovations.
 
Characters
 
Little mage is the main characters in this game. Mages are people who seek for knowledge and hidden secret of the world and practice sorcery.
 
Mage they can use their mana to create or produce spell element like fire, ice, and Etc. that they used to attack their enemies.
 
Story
In its earliest, and some would say most potent form, magic was primarily the art of memory. It required no technology, no wands or appurtenances other than the mind of the magician. All the trappings of ritual were merely mnemonic devices, meant to allow the practitioner to recall in rich detail the specific mental formulae that unlocked a spell's power. But some of the mage use their power to harm the other to dominate the world they called the black mage and this is the beginning of the journey of our little mage to stop the evil plan of black mage.
 
Theme
This is a game about the journey of the mage. Game theme is adventure and fantasy a mage go for a journey to stop the black mage in his journey he will encounter different mages that help him to enhance his knowledge/power.
 
Gameplay
Adventure and fantasy game. A journey of the Little mage to defeat the Black mage to bring back the peace and harmony to the world.
 
●     Goals: Defeat the Black Mage . Gameplay (short term): Defeat the enemies, advance to the next level, etc...
●     User Skills: Tap on the screen, Strategize
●     Items and power-ups (optional)
●     Progression and Challenge - Difficulty will advance by making the enemies harder.    	
●     Losing: These are the losing conditions: losing by lose your 3 life in the game.
 
Art Style
This is a 2D isometric game, with high quality 2D sprites. The character design should resemble that of Studio Ghibli. Everything should be very colorful and feel alive, with highly animated scenarios and layered backgrounds.
 
Music and Sounds
●     The music should have a Retro style, appealing to 8 bit nostalgia but high quality.
                                                                                	
●     It’s important that a lot of sound effects praise the user when he does something good. There should be immediate and positive feedback.
                                                                                	
                                                                                	
●     The sad scenes should be accompanied by Accordion/Violin music and sound like a sorrowful Tango.
                                                                                	
●     For In-Game music, use a more relaxed approach with happy tunes and going up on tempo as the level progresses. When in caves the music should sound muffled.

Technical Description
The game will be Desktop platform: With PC standalone version and Facebook Canvas.
 
●     Controls
 
●     Key-Arrow-Up to jump
●     Key-Arrow-Left to move backward
●     Key-Arrow-Right to move forward
●     Mouse-Leftclick to attack
 
●     Fonts
●     Audio / Music  
●     Graphics
1.	Characters
·         Little mage (idle, walking, running)
·         Normal mage (idle, walking, running).
·         Black mage (idle, walking, flying)
 
2.	Blocks
a.	Dirt
b.	Dirt/Grass
c. 	Stone Block
d.	Stone Bricks
 
3.	Ambient (Background)
a.	Tall Grass
b.	Old Trees
c. 	Torch
d.	Crow and bats
 
Demographics
Age: 12 to 50 , Sex: Everyone, Casual players mostly.
Platforms & Monetization
This game is Free to play with Ads.
 



Unsatisfactory - 2
Satisfactory - 3
Outstanding - 5
Documentation 
Some sections of the documents is not supplied and is incorrect format 
All sections of the documents are supplied and in correct format but not well elaborated 
All sections of the documents are complete, well  elaborated and in correct format. 
Gamename/Logo Banner 
Game name is not intuitive, colors and fonts do not create a single entity. 
Game name is not intuitive, colors and fonts create a single entity. 
Game name is unique and intuitive, colors and fonts create a single entity. 
Graphics
NO OTHER GRAPHICS SUPPLIED
Time Management
Document is submitted on time but game is not playable and some areas needs improvement
Document is submitted on time, game is playable but some areas needs improvement
Document is submitted on time, game is completely playable with complete assets
TOTAL SCORE
15